requirements.txt
flask
werkzeug
flask-jwt-extended
psycopg2-binary
python-dotenv
PyJWT==2.9.0
gunicorn
jsonify
kafka-python